+++ lua/nvim-yati/internal.lua
@@ -3,7 +3,7 @@ local stored_expr = {}
 
 function M.attach(bufnr, lang)
   stored_expr[bufnr] = vim.api.nvim_buf_get_option(bufnr, "indentexpr")
-  vim.api.nvim_buf_set_option(bufnr, "indentexpr", 'v:lua.require("nvim-yati.indent2").indentexpr()')
+  vim.api.nvim_buf_set_option(bufnr, "indentexpr", 'v:lua.require("nvim-yati.indent").indentexpr()')
 end
 
 function M.detach(bufnr)
